+++ CHANGELOG.md
@@ -4,4 +4,18 @@ This project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 
+## [0.0.4] -- 2019-02-26
+
+- Add some basic end-to-end tests.
+- Add additional copy using base64 encoding, by default.
+- Add samples and more README notes.
+
+## [0.0.3] -- 2018-08-15
+
+- Add option to change sixword font size.
+
+## [0.0.2] -- 2018-08-14
+
+- Initial public release
+
 <!-- vim: set tw=79 : -->